+++ src/client.js
@@ -129,30 +129,30 @@ export class Client extends EventEmitter {
    * @param {Configuration} config The parameters to define.
    */
   _updateJob(job, config) {
-    if (config.containsKey('repo_token')) job.repoToken = config.get('repo_token');
-    else if (config.containsKey('repo_secret_token')) job.repoToken = config.get('repo_secret_token');
+    if (config.has('repo_token')) job.repoToken = config.get('repo_token');
+    else if (config.has('repo_secret_token')) job.repoToken = config.get('repo_secret_token');
 
-    if (config.containsKey('parallel')) job.isParallel = config.get('parallel') == 'true';
-    if (config.containsKey('run_at')) job.runAt = new Date(config.get('run_at'));
-    if (config.containsKey('service_job_id')) job.serviceJobId = config.get('service_job_id');
-    if (config.containsKey('service_name')) job.serviceName = config.get('service_name');
-    if (config.containsKey('service_number')) job.serviceNumber = config.get('service_number');
-    if (config.containsKey('service_pull_request')) job.servicePullRequest = config.get('service_pull_request');
+    if (config.has('parallel')) job.isParallel = config.get('parallel') == 'true';
+    if (config.has('run_at')) job.runAt = new Date(config.get('run_at'));
+    if (config.has('service_job_id')) job.serviceJobId = config.get('service_job_id');
+    if (config.has('service_name')) job.serviceName = config.get('service_name');
+    if (config.has('service_number')) job.serviceNumber = config.get('service_number');
+    if (config.has('service_pull_request')) job.servicePullRequest = config.get('service_pull_request');
 
     let hasGitData = config.keys.some(key => key == 'service_branch' || key.substr(0, 4) == 'git_');
-    if (!hasGitData) job.commitSha = config.containsKey('commit_sha') ? config.get('commit_sha') : '';
+    if (!hasGitData) job.commitSha = config.has('commit_sha') ? config.get('commit_sha') : '';
     else {
       let commit = new GitCommit(
-        config.containsKey('commit_sha') ? config.get('commit_sha') : '',
-        config.containsKey('git_message') ? config.get('git_message') : ''
+        config.has('commit_sha') ? config.get('commit_sha') : '',
+        config.has('git_message') ? config.get('git_message') : ''
       );
 
-      commit.authorEmail = config.containsKey('git_author_email') ? config.get('git_author_email') : '';
-      commit.authorName = config.containsKey('git_author_name') ? config.get('git_author_name') : '';
-      commit.committerEmail = config.containsKey('git_committer_email') ? config.get('git_committer_email') : '';
-      commit.committerName = config.containsKey('git_committer_email') ? config.get('git_committer_email') : '';
+      commit.authorEmail = config.has('git_author_email') ? config.get('git_author_email') : '';
+      commit.authorName = config.has('git_author_name') ? config.get('git_author_name') : '';
+      commit.committerEmail = config.has('git_committer_email') ? config.get('git_committer_email') : '';
+      commit.committerName = config.has('git_committer_email') ? config.get('git_committer_email') : '';
 
-      job.git = new GitData(commit, config.containsKey('service_branch') ? config.get('service_branch') : '');
+      job.git = new GitData(commit, config.has('service_branch') ? config.get('service_branch') : '');
     }
   }
 }